Rename SceneListActivity=>FilterListActivity
